+++ src/Zicht/Bundle/KeyValueBundle/KeyValueStorage/PredefinedKey.php
@@ -5,12 +5,14 @@
 
 namespace Zicht\Bundle\KeyValueBundle\KeyValueStorage;
 
+use Swaggest\JsonSchema\Schema;
+
 /**
  * Class PredefinedKey.
  *
  * Definition of a Predefined-key.
  */
-class PredefinedKey
+class PredefinedKey implements PredefinedKeyInterface
 {
     /**
      * Create a key.
@@ -31,26 +33,6 @@ class PredefinedKey
         $instance->setFriendlyName($friendlyName);
         $instance->setFormType($formType);
         $instance->setFormOptions($formOptions);
-        return $instance;
-    }
-
-    /**
-     * Create a key based on a json schema
-     * Values can be scalar or array.
-     *
-     * @param string $jsonSchemaFile
-     * @param mixed $value
-     * @param string|null $friendlyName
-     * @return PredefinedKey
-     */
-    public static function createJsonSchemaKey($jsonSchemaFile, $value = null, $friendlyName = null)
-    {
-        $instance = new self();
-        $instance->setKey(basename($jsonSchemaFile));
-        $instance->setValue($value);
-        $instance->setFriendlyName($friendlyName);
-        $instance->setFormType("zicht_json_schema_type");
-        $instance->setFormOptions(['json_schema_file' => $jsonSchemaFile]);
         return $instance;
     }
 
@@ -91,7 +73,7 @@ class PredefinedKey
     /**
      * PredefinedKey constructor.
      *
-     * Disable constructing, they can only be created from self::createKey to ensure key/value immutablilty.
+     * Disable constructing, they can only be created from self::createKey to ensure key/value immutability.
      */
     private function __construct()
     {
@@ -138,7 +120,7 @@ class PredefinedKey
     }
 
     /**
-     * @return string
+     * {@inheritDoc}
      */
     public function getKey()
     {
@@ -146,7 +128,7 @@ class PredefinedKey
     }
 
     /**
-     * @return mixed
+     * {@inheritDoc}
      */
     public function getValue()
     {
@@ -154,7 +136,7 @@ class PredefinedKey
     }
 
     /**
-     * @return string
+     * {@inheritDoc}
      */
     public function getFriendlyName()
     {
@@ -162,7 +144,7 @@ class PredefinedKey
     }
 
     /**
-     * @return string
+     * {@inheritDoc}
      */
     public function getFormType()
     {
@@ -170,7 +152,7 @@ class PredefinedKey
     }
 
     /**
-     * @return array
+     * {@inheritDoc}
      */
     public function getFormOptions()
     {