+++ src/Zicht/Bundle/KeyValueBundle/KeyValueStorage/PredefinedKeyInterface.php
@@ -1,0 +1,34 @@
+<?php
+/**
+ * @copyright Zicht Online <http://www.zicht.nl>
+ */
+
+namespace Zicht\Bundle\KeyValueBundle\KeyValueStorage;
+
+interface PredefinedKeyInterface
+{
+    /**
+     * @return string
+     */
+    public function getKey();
+
+    /**
+     * @return mixed
+     */
+    public function getValue();
+
+    /**
+     * @return string
+     */
+    public function getFriendlyName();
+
+    /**
+     * @return string
+     */
+    public function getFormType();
+
+    /**
+     * @return array
+     */
+    public function getFormOptions();
+}